# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maktab Raqamli Tizimi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /grades:
    get:
      operationId: listGrades
      tags: [grades]
      summary: List grades
      parameters:
        - in: query
          name: studentId
          schema:
            type: number
        - in: query
          name: classId
          schema:
            type: number
        - in: query
          name: subjectId
          schema:
            type: number
        - in: query
          name: teacherId
          schema:
            type: number
      responses:
        "200":
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
    post:
      operationId: createGrade
      tags: [grades]
      summary: Create grade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /grades/{id}:
    get:
      operationId: getGrade
      tags: [grades]
      summary: Get grade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Grade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
    patch:
      operationId: updateGrade
      tags: [grades]
      summary: Update grade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
    delete:
      operationId: deleteGrade
      tags: [grades]
      summary: Delete grade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /dashboard/weekly-summary:
    get:
      operationId: getWeeklySummary
      tags: [dashboard]
      summary: Get weekly AI summary for parent/student
      parameters:
        - in: query
          name: studentId
          schema:
            type: number
      responses:
        "200":
          description: Weekly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklySummary"
  /dashboard/grade-stats:
    get:
      operationId: getGradeStats
      tags: [dashboard]
      summary: Get grade statistics per class/subject
      parameters:
        - in: query
          name: classId
          schema:
            type: number
        - in: query
          name: subjectId
          schema:
            type: number
      responses:
        "200":
          description: Grade stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GradeStat"
components:
  schemas:
...
    Class:
      type: object
      required: [id, name, grade, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        grade:
          type: number
        teacherId:
          type: ["integer", "null"]
        studentCount:
          type: number
        createdAt:
          type: string
          format: date-time
    ClassInput:
      type: object
      required: [name, grade]
      properties:
        name:
          type: string
        grade:
          type: number
        teacherId:
          type: number
    ClassUpdate:
      type: object
      properties:
        name:
          type: string
        grade:
          type: number
        teacherId:
          type: number
...
    Grade:
      type: object
      required: [id, studentId, subjectId, teacherId, value, type, date, createdAt]
      properties:
        id:
          type: number
        studentId:
          type: number
        studentName:
          type: string
        subjectId:
          type: number
        subjectName:
          type: string
        teacherId:
          type: number
        classId:
          type: ["integer", "null"]
        value:
          type: number
          minimum: 1
          maximum: 5
        type:
          type: string
          enum: [daily, quarter, annual, test]
        comment:
          type: ["string", "null"]
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    GradeInput:
      type: object
      required: [studentId, subjectId, teacherId, value, type, date]
      properties:
        studentId:
          type: number
        subjectId:
          type: number
        teacherId:
          type: number
        classId:
          type: number
        value:
          type: number
          minimum: 1
          maximum: 5
        type:
          type: string
          enum: [daily, quarter, annual, test]
        comment:
          type: string
        date:
          type: string
          format: date
    GradeUpdate:
      type: object
      properties:
        value:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
...
    WeeklySummary:
      type: object
      required: [studentId, studentName, weekStart, weekEnd, summary, gradesThisWeek, absencesThisWeek, assignmentsDue]
      properties:
        studentId:
          type: number
        studentName:
          type: string
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        summary:
          type: string
        gradesThisWeek:
          type: number
        absencesThisWeek:
          type: number
        assignmentsDue:
          type: number
        averageGrade:
          type: number
        subjectHighlights:
          type: array
          items:
            $ref: "#/components/schemas/SubjectHighlight"
    SubjectHighlight:
      type: object
      required: [subjectName, status, note]
      properties:
        subjectName:
          type: string
        status:
          type: string
          enum: [improving, stable, declining]
        note:
          type: string
    GradeStat:
      type: object
      required: [label, average, count]
      properties:
        label:
          type: string
        average:
          type: number
        count:
          type: number
        distribution:
          type: array
          items:
            $ref: "#/components/schemas/GradeDistributionItem"
    GradeDistributionItem:
      type: object
      required: [value, count]
      properties:
        value:
          type: number
        count:
          type: number